+++ crates/wdl-engine/src/eval/v1/task.rs
@@ -595,7 +595,9 @@ impl Evaluator {
             .perform_task_evaluation(document, task, inputs, eval_root_dir.as_ref(), task.name())
             .await;
 
-        if self.cancellation.user_canceled() {
+        if self.cancellation.user_canceled()
+            && self.cancellation.state() == CancellationContextState::Canceling
+        {
             return Err(EvaluationError::Canceled);
         }